Almost there
Ready: Register, login, home page views based on login status, sorting of plays that display - only the 3 most liked plays are displayed for unauthenticated users. Logged in users see all the plays arranged by their date of creation.

To do: Details page > options to edit, delete and like the play (for users who are not owners of the play)
Options to sort the plays in the home page by number of likes or date created.